# swagger.yaml
openapi: 3.0.3
info:
  title: VBI-Learning API Documentation
  description: API documentation for VBI-Learning platform
  version: 1.0.0
servers:
  - url: https://vbi-learning-api.example.com
...
paths:
...
  /mentor/create:
    post:
      tags:
        - Mentor
      summary: Create a new mentor
      operationId: createMentor
      requestBody:
        description: Mentor data to be created
        required: true
        content:
          application/json:
            example:
              mentoring_title: 'Mentoring Title'
              mentoring_owner: 'Mentor Owner'
      responses:
        '201':
          description: Mentor created successfully
          content:
            application/json:
              example: { /* Mentor details */ }
        '500':
          description: Failed to create mentor
          content:
            application/json:
              example: { error: 'Failed to create mentor' }
  /mentor/update/{mentorId}:
    put:
      tags:
        - Mentor
      summary: Update an existing mentor
      operationId: updateMentor
      parameters:
        - name: mentorId
          in: path
          required: true
          schema:
            type: string
          description: ID of the mentor to be updated
      requestBody:
        description: Mentor data to be updated
        required: true
        content:
          application/json:
            example:
              mentoring_title: 'Updated Mentoring Title'
              mentoring_owner: 'Updated Mentor Owner'
      responses:
        '200':
          description: Mentor updated successfully
          content:
            application/json:
              example: { /* Updated mentor details */ }
        '500':
          description: Failed to update mentor
          content:
            application/json:
              example: { error: 'Failed to update mentor' }
  /mentor/delete/{mentorId}:
    delete:
      tags:
        - Mentor
      summary: Delete a mentor
      operationId: deleteMentor
      parameters:
        - name: mentorId
          in: path
          required: true
          schema:
            type: string
          description: ID of the mentor to be deleted
      responses:
        '200':
          description: Mentor deleted successfully
          content:
            application/json:
              example: { message: 'Mentor deleted successfully' }
        '500':
          description: Failed to delete mentor
          content:
            application/json:
              example: { error: 'Failed to delete mentor' }
  /mentor/{mentorId}:
    get:
      tags:
        - Mentor
      summary: Get mentor details by ID
      operationId: getMentorById
      parameters:
        - name: mentorId
          in: path
          required: true
          schema:
            type: string
          description: ID of the mentor to fetch details
      responses:
        '200':
          description: Mentor details
          content:
            application/json:
              example: { /* Mentor details */ }
        '500':
          description: Failed to get mentor
          content:
            application/json:
              example: { error: 'Failed to get mentor' }
  /mentors:
    get:
      tags:
        - Mentor
      summary: Get details of all mentors
      operationId: getAllMentors
      responses:
        '200':
          description: List of all mentors
          content:
            application/json:
              example: [{ /* Mentor details */ }, { /* Mentor details */ }]
        '500':
          description: Failed to get mentors
          content:
            application/json:
              example: { error: 'Failed to get mentors' }
  /pool/create:
    post:
      tags:
        - Pool
      summary: Create a new pool
      operationId: createPool
      requestBody:
        description: Pool data to be created
        required: true
        content:
          application/json:
            example:
              pool_id: 'pool123'
              owner_id: 'user123'
              minimum_stake: 10
              maximum_stake: 100
              staking_period: 30
              unstaking_period: 15
              description: 'Sample pool description'
              image: 'pool_image_url'
      responses:
        '201':
          description: Pool created successfully
          content:
            application/json:
              example: { /* Created pool details */ }
        '500':
          description: Failed to create pool
          content:
            application/json:
              example: { error: 'Failed to create pool' }
  /pool/update/{poolId}:
    put:
      tags:
        - Pool
      summary: Update an existing pool
      operationId: updatePool
      parameters:
        - name: poolId
          in: path
          required: true
          schema:
            type: string
          description: ID of the pool to be updated
      requestBody:
        description: Pool data to be updated
        required: true
        content:
          application/json:
            example: /* Updated pool data example */
      responses:
        '200':
          description: Pool updated successfully
          content:
            application/json:
              example: { /* Updated pool details */ }
        '404':
          description: Pool not found
          content:
            application/json:
              example: { error: 'Pool not found' }
        '500':
          description: Failed to update pool
          content:
            application/json:
              example: { error: 'Failed to update pool' }
  /pool/delete/{poolId}:
    delete:
      tags:
        - Pool
      summary: Delete a pool
      operationId: deletePool
      parameters:
        - name: poolId
          in: path
          required: true
          schema:
            type: string
          description: ID of the pool to be deleted
      responses:
        '200':
          description: Pool deleted successfully
          content:
            application/json:
              example: { message: 'Pool deleted successfully' }
        '404':
          description: Pool not found
          content:
            application/json:
              example: { error: 'Pool not found' }
        '500':
          description: Failed to delete pool
          content:
            application/json:
              example: { error: 'Failed to delete pool' }
  /pool/{poolId}:
    get:
      tags:
        - Pool
      summary: Get pool details by ID
      operationId: getByPoolId
      parameters:
        - name: poolId
          in: path
          required: true
          schema:
            type: string
          description: ID of the pool to fetch details
      responses:
        '200':
          description: Pool details
          content:
            application/json:
              example: { /* Pool details */ }
        '404':
          description: Pool not found
          content:
            application/json:
              example: { error: 'Pool not found' }
        '500':
          description: Failed to get pool
          content:
            application/json:
              example: { error: 'Failed to get pool by pool_id' }
  /pool:
    get:
      tags:
        - Pool
      summary: Get details of all pools
      operationId: getAllPools
      responses:
        '200':
          description: List of all pools
          content:
            application/json:
              example: [{ /* Pool details */ }, { /* Pool details */ }]
        '500':
          description: Failed to get all pools
          content:
            application/json:
              example: { error: 'Failed to get all pools' }